Add files via upload
Application update - Fixed data filtering when the main data-table page is loaded. Added more datatable columns (Payment due and Total cost). Added automatic writing for the input of the email. Added filter in the main datatable page to filter data if the invoice is issued or not. Added the same options for the details page - to issue invoices from there.